get messages is broken
Check in to get some assistance with my getMessages API call.  Tried various ways of attaching headers/parameters.  At this point the call coming from HttpClient is failing or being blocked, but when I log the request in the console I can copy the urlWithParams into a browser and get the response back from the server.  What is my HttpClient request missing?